+++ packages/web-components/src/components/dotcom-shell/dotcom-shell-composite.ts
@@ -25,13 +25,13 @@ import { UNAUTHENTICATED_STATUS } from '../../internal/vendor/@carbon/ibmdotcom-
 import { FOOTER_SIZE } from '../footer/footer';
 import '../footer/footer-composite';
 import './dotcom-shell';
-import styles from './dotcom-shell-composite.scss';
+import styles from './dotcom-shell.scss';
 import { carbonElement as customElement } from '../../internal/vendor/@carbon/web-components/globals/decorators/carbon-element';
 
 const { stablePrefix: c4dPrefix } = settings;
 
 /**
- * Component that rendres dotcom shell from links, etc. data.
+ * Component that renders dotcom shell from links, etc. data.
  *
  * @element c4d-dotcom-shell-composite
  */
@@ -288,12 +288,6 @@ class C4DDotcomShellComposite extends LitElement {
   selectedMenuItem!: string;
 
   /**
-   * `true` to open the locale modal. This goes to footer.
-   */
-  @property({ type: Boolean, attribute: 'open-locale-modal' })
-  openLocaleModal = false;
-
-  /**
    * Footer size. This goes to footer.
    */
   @property({ reflect: true, attribute: 'footer-size' })
@@ -386,7 +380,6 @@ class C4DDotcomShellComposite extends LitElement {
       localeList,
       footerLinks,
       footerSize,
-      openLocaleModal,
       openSearchDropdown,
       navLinks,
       hasProfile,
@@ -453,7 +446,6 @@ class C4DDotcomShellComposite extends LitElement {
           legalLinks,
           links: footerLinks,
           localeList,
-          openLocaleModal,
           selectedLanguage,
           size: footerSize,
           _loadLocaleList,